requirements.txt
aiohttp >= "3.7"
pytz >= "2021.1"
discord.py == "2.0.0"
bs4 >= "0.0.1"
yt-dlp >= "2023.2.17"
PyNaCl >= "1.4.0"
google-api-python-client >= "2.21.0"
spotipy >= "2.19.0"
psutil >= "5.8.0"
redis >= "4.0.0"
Pillow >= "1.0